+++ pkg/ctl/bk/ledger/delete.go
