+++ pkg/pulsarctl.go
@@ -19,7 +19,7 @@ package pkg
 
 import (
 	"github.com/streamnative/pulsarctl/pkg/cmdutils"
-	"github.com/streamnative/pulsarctl/pkg/ctl/bookie"
+	"github.com/streamnative/pulsarctl/pkg/ctl/bk"
 	"github.com/streamnative/pulsarctl/pkg/ctl/cluster"
 	"github.com/streamnative/pulsarctl/pkg/ctl/completion"
 	"github.com/streamnative/pulsarctl/pkg/ctl/namespace"
@@ -88,7 +88,7 @@ func NewPulsarctlCmd() *cobra.Command {
 	rootCmd.AddCommand(topic.Command(flagGrouping))
 	rootCmd.AddCommand(namespace.Command(flagGrouping))
 	rootCmd.AddCommand(schema.Command(flagGrouping))
-	rootCmd.AddCommand(bookie.Command(flagGrouping))
+	rootCmd.AddCommand(bk.Command(flagGrouping))
 	rootCmd.AddCommand(subscription.Command(flagGrouping))
 
 	return rootCmd
